apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
httpx==0.28.1
beautifulsoup4==4.13.4
lxml==6.0.0
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
redis==6.2.0
celery==5.5.3
pytest==8.4.1

